# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Buc", "version": "0.1.0" },
  "paths": {
    "/api/v1/login/access-token": {
      "post": {
        "tags": ["login"],
        "summary": "Login Access Token",
        "description": "OAuth2 compatible token login, get an access token for future requests",
        "operationId": "login-login_access_token",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": { "$ref": "#/components/schemas/Body_login-login_access_token" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Token" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/login/test-token": {
      "post": {
        "tags": ["login"],
        "summary": "Test Token",
        "description": "Test access token",
        "operationId": "login-test_token",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/password-recovery/{email}": {
      "post": {
        "tags": ["login"],
        "summary": "Recover Password",
        "description": "Password Recovery",
        "operationId": "login-recover_password",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Email" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/reset-password/": {
      "post": {
        "tags": ["login"],
        "summary": "Reset Password",
        "description": "Reset password",
        "operationId": "login-reset_password",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/NewPassword" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/password-recovery-html-content/{email}": {
      "post": {
        "tags": ["login"],
        "summary": "Recover Password Html Content",
        "description": "HTML Content for Password Recovery",
        "operationId": "login-recover_password_html_content",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Email" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "text/html": { "schema": { "type": "string" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/companies/": {
      "get": {
        "tags": ["companies"],
        "summary": "Read Companies",
        "description": "Retrieve companies.",
        "operationId": "companies-read_companies",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CompaniesPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["companies"],
        "summary": "Create Company",
        "description": "Create new company.",
        "operationId": "companies-create_company",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CompanyCreate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CompanyPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{company_id}": {
      "get": {
        "tags": ["companies"],
        "summary": "Read Company By Id",
        "description": "Get a specific company by id.",
        "operationId": "companies-read_company_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Company Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CompanyPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["companies"],
        "summary": "Update Company",
        "description": "Update a company.",
        "operationId": "companies-update_company",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Company Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CompanyUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CompanyPublic" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["companies"],
        "summary": "Delete Company",
        "description": "Delete a company.",
        "operationId": "companies-delete_company",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Company Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_login-login_access_token": {
        "properties": {
          "grant_type": {
            "anyOf": [{ "type": "string", "pattern": "password" }, { "type": "null" }],
            "title": "Grant Type"
          },
          "username": { "type": "string", "title": "Username" },
          "password": { "type": "string", "title": "Password" },
          "scope": { "type": "string", "title": "Scope", "default": "" },
          "client_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": ["username", "password"],
        "title": "Body_login-login_access_token"
      },
      "CompaniesPublic": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/CompanyPublic" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "CompaniesPublic"
      },
      "CompanyCreate": {
        "properties": {
          "name": { "type": "string", "maxLength": 255, "title": "Name" },
          "details": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Details" },
          "is_demo": { "type": "boolean", "title": "Is Demo", "default": false }
        },
        "type": "object",
        "required": ["name"],
        "title": "CompanyCreate"
      },
      "CompanyPublic": {
        "properties": {
          "name": { "type": "string", "maxLength": 255, "title": "Name" },
          "details": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Details" },
          "is_demo": { "type": "boolean", "title": "Is Demo", "default": false },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "created_at": { "type": "string", "format": "date-time", "title": "Created At" },
          "updated_at": { "type": "string", "format": "date-time", "title": "Updated At" }
        },
        "type": "object",
        "required": ["name", "id", "created_at", "updated_at"],
        "title": "CompanyPublic"
      },
      "CompanyUpdate": {
        "properties": {
          "name": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Name" },
          "details": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Details" },
          "is_demo": { "anyOf": [{ "type": "boolean" }, { "type": "null" }], "title": "Is Demo" }
        },
        "type": "object",
        "title": "CompanyUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": { "message": { "type": "string", "title": "Message" } },
        "type": "object",
        "required": ["message"],
        "title": "Message"
      },
      "NewPassword": {
        "properties": {
          "token": { "type": "string", "title": "Token" },
          "new_password": { "type": "string", "title": "New Password" }
        },
        "type": "object",
        "required": ["token", "new_password"],
        "title": "NewPassword"
      },
...
      "Token": {
        "properties": {
          "access_token": { "type": "string", "title": "Access Token" },
          "token_type": { "type": "string", "title": "Token Type", "default": "bearer" }
        },
        "type": "object",
        "required": ["access_token"],
        "title": "Token"
      },
...
      "UserPublic": {
        "properties": {
          "email": { "type": "string", "maxLength": 255, "title": "Email" },
          "is_active": { "type": "boolean", "title": "Is Active", "default": true },
          "is_superuser": { "type": "boolean", "title": "Is Superuser", "default": false },
          "full_name": {
            "anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
            "title": "Full Name"
          },
          "role": { "$ref": "#/components/schemas/UserRole", "default": "USER" },
          "is_verified": { "type": "boolean", "title": "Is Verified", "default": false },
          "company_id": {
            "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }],
            "title": "Company Id"
          },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "created_at": { "type": "string", "format": "date-time", "title": "Created At" }
        },
        "type": "object",
        "required": ["email", "id", "created_at"],
        "title": "UserPublic"
      },
...
      "UserRole": { "type": "string", "enum": ["ADMIN", "AUDITOR", "USER"], "title": "UserRole" },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" } }
      }
    }
  }
}